server/requirements.txt
flask
config
flask_cors
mysql-connector-python
transliterate